removed roles and added it to second page
